test: split custom-parser.test.js
